Refactor carpeta implementaciones
