--- Clase-3-MVC/Clase-3-MVC.Web/Servicio/IServicioPartido.cs
@@ -1,13 +1,0 @@
-﻿using Clase_3_MVC.Web.Models;
-using Microsoft.AspNetCore.Http;
-using System.Collections.Generic;
-
-namespace Clase_3_MVC.Web.Servicio
-{
-    public interface IServicioPartido
-    {
-        List<PartidoViewModel> ObtenerPartidos();
-        List<PartidoViewModel> ConsultarFecha(IFormCollection colection);
-        void AgregarNuevoPartido(IFormCollection collection);
-    }
-}
